enable printf-style argument checking on OS X which (unfortunately) requires adding explicit casts every time a size_t value is output since the type is 'long unsigned int' on OS X, as opposed to 'unsigned int' on Linux and others

+++ debug/Log.cc
@@ -249,7 +249,7 @@ Log::parse_debug_file(const char* debug_path)
     rule_list_ = new_rule_list;
     if (inited_) {
         logf("/log", LOG_ALWAYS, "reparsed debug file... found %d rules",
-             new_rule_list->size());
+             (int)new_rule_list->size());
     }
     lock_->unlock();
